refactor: move BudgetInterface to independent file
