+++ budget_service.py
@@ -4,23 +4,7 @@ from calendar import monthrange
 from typing import List
 from dateutil.relativedelta import relativedelta
 
-
-class Budget:
-    def __init__(self, year_month, amount):
-        self.year_month = year_month
-        self.amount = amount
-
-
-class BudgetsInterface:
-    budgets = []
-
-    @classmethod
-    def get_all(cls) -> List[Budget]:
-        return cls.budgets
-
-    @classmethod
-    def inject_mock_data(cls, budgets) -> None:
-        cls.budgets = budgets
+from budget_interface import Budget, BudgetsInterface
 
 
 class BudgetService:
